[Issue Tracker] Minor cleanup (#2751)
Breakdown react components in separate files + CSS
