+++ modules/issue_tracker/php/NDB_Form_issue_tracker.class.inc
@@ -100,7 +100,7 @@ class NDB_Form_Issue_Tracker extends NDB_Form
         $deps    = parent::getJSDependencies();
         return array_merge(
             $deps,
-            [$baseURL . "/issue_tracker/js/issueIndex.js"]
+            [$baseURL . "/issue_tracker/js/index.js"]
         );
     }
